+++ CHANGELOG.md
@@ -3,10 +3,12 @@
 ## [Unreleased]
 
 * Update PowerSTIG to successfully parse/apply Microsoft Windows 2012 R2 DC STIG- Ver 3, Rel 2 [#902](https://github.com/microsoft/PowerStig/issues/902)
+* Added SQL Server 2016 vulnerability ID's: [#872](https://github.com/microsoft/PowerStig/issues/872)
 * Update PowerSTIG to remove old rule Ids in Hard Coded Framework: [#864](https://github.com/microsoft/PowerStig/issues/864)
 * Fixed: PowerShell v7 Fix: AccessControlDsc/RegistryAccessEntry & NTFSAccessEntry "AccessControlList" should be an Array
 * Update PowerSTIG to Parse/Apply Google Chrome Ver 2, Rel 2: [#876](https://github.com/microsoft/PowerStig/issues/876)
 * Fix Validation test fail because TAB's are found in Description Field [#878](https://github.com/microsoft/PowerStig/issues/878)
+
 
 ## [4.8.0] - 2021-03-01